Fix runtime env var loading
- Add config:clear to entrypoint to ensure runtime env vars are used
- Add int casts for numeric config values

+++ config/reverb.php
@@ -19,17 +19,17 @@ return [
     'servers' => [
         'reverb' => [
             'host' => env('REVERB_HOST', '0.0.0.0'),
-            'port' => env('REVERB_PORT', 8080),
+            'port' => (int) env('REVERB_PORT', 8080),
             'hostname' => env('REVERB_HOST', '0.0.0.0'),
             'options' => [
                 'tls' => [],
             ],
-            'max_request_size' => env('REVERB_MAX_REQUEST_SIZE', 10_000),
+            'max_request_size' => (int) env('REVERB_MAX_REQUEST_SIZE', 10_000),
             'scaling' => [
                 'enabled' => false,
             ],
-            'pulse_ingest_interval' => env('REVERB_PULSE_INGEST_INTERVAL', 15),
-            'telescope_ingest_interval' => env('REVERB_TELESCOPE_INGEST_INTERVAL', 15),
+            'pulse_ingest_interval' => (int) env('REVERB_PULSE_INGEST_INTERVAL', 15),
+            'telescope_ingest_interval' => (int) env('REVERB_TELESCOPE_INGEST_INTERVAL', 15),
         ],
     ],
 
@@ -46,14 +46,14 @@ return [
             'app_id' => env('REVERB_APP_ID', 'app-id'),
             'options' => [
                 'host' => env('REVERB_HOST', '0.0.0.0'),
-                'port' => env('REVERB_PORT', 8080),
+                'port' => (int) env('REVERB_PORT', 8080),
                 'scheme' => env('REVERB_SCHEME', 'http'),
                 'useTLS' => env('REVERB_SCHEME', 'http') === 'https',
             ],
             'allowed_origins' => ['*'],
-            'ping_interval' => env('REVERB_APP_PING_INTERVAL', 60),
-            'activity_timeout' => env('REVERB_APP_ACTIVITY_TIMEOUT', 30),
-            'max_message_size' => env('REVERB_APP_MAX_MESSAGE_SIZE', 10_000),
+            'ping_interval' => (int) env('REVERB_APP_PING_INTERVAL', 60),
+            'activity_timeout' => (int) env('REVERB_APP_ACTIVITY_TIMEOUT', 30),
+            'max_message_size' => (int) env('REVERB_APP_MAX_MESSAGE_SIZE', 10_000),
         ],
     ],
